+++ test/interfaces/v8/IPushMock.sol
@@ -6,8 +6,8 @@ pragma solidity 0.8.29;
  * @dev Minimal interface needed for MigrationLocker contract tests
  */
 interface IPushMock {
-    function transferFrom(address src, address dst, uint rawAmount) external returns (bool);
+    function transferFrom(address src, address dst, uint256 rawAmount) external returns (bool);
     function burn(uint256 rawAmount) external;
-    function balanceOf(address account) external view returns (uint);
-    function transfer(address dst, uint rawAmount) external returns (bool);
-} 
+    function balanceOf(address account) external view returns (uint256);
+    function transfer(address dst, uint256 rawAmount) external returns (bool);
+}
